Main.javaの作成。 classの分類。
